+++ src/solve/Matrix/frobenius_product_symmetric.cxx
@@ -8,14 +8,14 @@ Real frobenius_product_symmetric(const Matrix &A, const Matrix &B)
   assert(A.rows == A.cols);
 
   Real result = 0;
-  for(int c = 0; c < A.cols; c++)
-    for(int r = 0; r < c; r++)
+  for(size_t c = 0; c < A.cols; c++)
+    for(size_t r = 0; r < c; r++)
       {
         result += A.elt(r, c) * B.elt(r, c);
       }
   result *= 2;
 
-  for(int r = 0; r < A.rows; r++)
+  for(size_t r = 0; r < A.rows; r++)
     {
       result += A.elt(r, r) * B.elt(r, r);
     }
