+++ src/solve/Matrix/lower_triangular_solve.cxx
@@ -4,7 +4,7 @@
 // pointed to by b
 void lower_triangular_solve(const Matrix &L, Real *b, int bcols, int ldb)
 {
-  int dim = L.rows;
+  size_t dim = L.rows;
   assert(L.cols == dim);
 
   Rtrsm("Left", "Lower", "NoTranspose", "NonUnitDiagonal", dim, bcols, 1,
@@ -14,6 +14,6 @@ void lower_triangular_solve(const Matrix &L, Real *b, int bcols, int ldb)
 // b := L^{-1} b, where L is lower-triangular
 void lower_triangular_solve(const Matrix &L, Vector &b)
 {
-  assert(static_cast<int>(b.size()) == L.rows);
+  assert(b.size() == L.rows);
   lower_triangular_solve(L, b.data(), 1, b.size());
 }
