+++ src/solve/Matrix/lower_triangular_transpose_solve.cxx
@@ -5,7 +5,7 @@
 void lower_triangular_transpose_solve(const Matrix &L, Real *b, int bcols,
                                       int ldb)
 {
-  int dim = L.rows;
+  size_t dim = L.rows;
   assert(L.cols == dim);
 
   Rtrsm("Left", "Lower", "Transpose", "NonUnitDiagonal", dim, bcols, 1,
@@ -15,6 +15,6 @@ void lower_triangular_transpose_solve(const Matrix &L, Real *b, int bcols,
 // b := L^{-T} b, where L is lower-triangular
 void lower_triangular_transpose_solve(const Matrix &L, Vector &b)
 {
-  assert(static_cast<int>(b.size()) == L.rows);
+  assert(b.size() == L.rows);
   lower_triangular_transpose_solve(L, b.data(), 1, b.size());
 }
