+++ src/solve/Matrix/matrix_solve_with_cholesky.cxx
@@ -3,7 +3,7 @@
 // X := ACholesky^{-1 T} ACholesky^{-1} X = A^{-1} X
 void matrix_solve_with_cholesky(const Matrix &ACholesky, Matrix &X)
 {
-  int dim = X.rows;
+  size_t dim = X.rows;
   assert(X.cols == dim);
   assert(ACholesky.rows == dim);
   assert(ACholesky.cols == dim);
